grid on gpu, with error

+++ gpu/struct.h
@@ -261,10 +261,10 @@ __device__ Color Shader_triangle(ShapeEntity *shape, void *pos, void *normal, Co
     float w = 1.0f - u - v;
 
     UVPoint mp = Vector(tri.t1).scale(v) + Vector(tri.t2).scale(w) + Vector(tri.t3).scale(u);
-    int x = tri.tex->width * mp.x;
-    int y = tri.tex->height * mp.y;
-    x = max(0, min(tri.tex->width - 1, x));
-    y = max(0, min(tri.tex->height - 1, y));
+    size_t x = tri.tex->width * mp.x;
+    size_t y = tri.tex->height * mp.y;
+    x = max((size_t)0, min(tri.tex->width - 1, x));
+    y = max((size_t)0, min(tri.tex->height - 1, y));
     return Vector::Mul(tri.tex->tex[tri.tex->width * y + x].v, factor->v);
 }
 typedef Color(*shader_fun_t)(ShapeEntity *, void *, void *, Color *);
